Exercises 4.15.-4.23. done including starred ones

+++ part4-testing-express-user-auth/blog-list/controllers/blogs.js
@@ -1,28 +1,74 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
+const middleware = require('../utils/middleware');
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogsRouter.post('/', async (request, response) => {
-  const { title, url } = request.body;
+blogsRouter.post(
+  '/',
+  middleware.tokenExtractor,
+  middleware.userExtractor,
+  async (request, response, next) => {
+    const { title, url } = request.body;
+    const userId = request.user.id;
 
-  if (!title || !url) {
-    response.status(400).end();
-  } else {
-    const blog = new Blog(request.body);
+    try {
+      const user = await User.findById(userId);
 
-    const result = await blog.save();
-    response.status(201).json(result);
+      if (!title || !url) {
+        response.status(400).end();
+      } else {
+        const blog = new Blog({
+          ...request.body,
+          user: userId,
+        });
+
+        user.blogs = user.blogs.concat(blog._id);
+
+        const savedBlog = await blog.save();
+
+        user.save();
+
+        response.status(201).json(savedBlog);
+      }
+    } catch (error) {
+      next(error);
+    }
   }
-});
+);
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndDelete(request.params.id);
-  result ? response.status(204).json(result) : response.status(404).end();
-});
+blogsRouter.delete(
+  '/:id',
+  middleware.tokenExtractor,
+  middleware.userExtractor,
+  async (request, response, next) => {
+    try {
+      const user = request.user;
+      const blog = await Blog.findById(request.params.id);
+
+      if (blog) {
+        if (blog.user.toString() === user.id) {
+          const res = await Blog.findByIdAndDelete(request.params.id);
+          response.status(204).json(res);
+        } else {
+          response
+            .status(400)
+            .json({ error: "The blog's user id and token id cannot match" });
+        }
+      } else {
+        response
+          .status(404)
+          .json({ error: 'Blog with given id cannot be found' });
+      }
+    } catch (error) {
+      next(error);
+    }
+  }
+);
 
 blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body;
